feat: firebase claims, addDate, deleteDate, tests

+++ src/controllers/User.js
@@ -1,6 +1,6 @@
 // import models
 const UserModel = require('../models/User');
-
+const DoctorModel = require('../models/Doctor');
 
 /**
  * Definições públicas
@@ -9,7 +9,7 @@ module.exports = {
     async create(req, res) {
         const { name, cpf, age, phone, firebaseUID } = req.body
     
-        if(name == null || cpf == null || age == null || phone == null, firebaseUID == null)
+        if(name == null || cpf == null || age == null || phone == null || firebaseUID == null)
             return res.status(206).send({ error: 'Insufficient data' });
 
         try {
@@ -45,6 +45,10 @@ module.exports = {
     async delete(req, res) {
         const {firebaseUID} = req.body;
 
+            
+        if(firebaseUID == null)
+            return res.status(206).send({ error: 'Insufficient data' });
+
         try {
             const response = await UserModel.deleteOne({ firebaseUID : firebaseUID });
 
@@ -54,6 +58,40 @@ module.exports = {
             return res.status(400).send({ error: 'Problem when deleting user' });
         } catch (err) {
             console.log(err);
+            return res.status(500).send({ error: 'Server request failed' });
+        }
+    },
+
+    async markCalendar(req, res) {
+        const { date, crm, firebaseUID } = req.body;
+
+            
+        if(date == null || crm == null || firebaseUID == null)
+            return res.status(206).send({ error: 'Insufficient data' });
+
+        if(Object.prototype.toString.call(schedule) === '[object Date]')
+            return res.status(400).send({ error: 'Invalid type date' });
+
+        try {
+            const doctor = DoctorModel.findOne({crm : crm});
+            
+            doctor.schedule.map((data) => {
+                if(data.year == date.year && data.month == date.month && data.day == date.day && 
+                    data.hour == date.hour && data.minute == date.minute) {
+                        return res.status(400).send({ 
+                            error: 'Doctor already has an appointment at that time'
+                        });
+                }
+            })
+
+            if(doctor){
+                doctor.schedule.push(date)
+                doctor.save()
+            }
+
+            return res.status(200).send({ message: 'Saved medical consultation' });
+
+        } catch (err) {
             return res.status(500).send({ error: 'Server request failed' });
         }
     }
